[reggen] Refactor reggen, add "local" to param
- Moved data structure into `util/reggen/data.py`
- add `local` field in `param_list`
  + if `local` is false, it creates parameter in reg_top.sv
- Revised HMAC, RV_PLIC to use parameter
