+++ hw/ip/rv_plic/doc/rv_plic.tpl.hjson
@@ -13,12 +13,26 @@
   clock_primary: "clk_fixed",
   bus_device: "tlul",
 
+  param_list: [
+    { name: "NumSrc",
+      desc: "Number of interrupt sources",
+      type: "int",
+      default: "${src}",
+      local: "true"
+    },
+    { name: "NumTarget",
+      desc: "Number of Targets (Harts)",
+      type: "int",
+      default: "${target}",
+      local: "true",
+    },
+  ],
   regwidth: "32",
   registers: [
     { multireg: {
         name: "IP",
         desc: "Interrupt Pending",
-        count: ${src},
+        count: "NumSrc",
         cname: "RV_PLIC",
         swaccess: "ro",
         hwaccess: "hwo",
@@ -30,7 +44,7 @@
     { multireg: {
         name: "LE",
         desc: "Interrupt Source mode. 0: Level, 1: Edge-triggered",
-        count: ${src},
+        count: "NumSrc",
         cname: "RV_PLIC",
         swaccess: "rw",
         hwaccess: "hro",
@@ -54,7 +68,7 @@
     { multireg: {
         name: "IE${i}",
         desc: "Interrupt Enable for Target ${i}",
-        count: ${src},
+        count: "NumSrc",
         cname: "RV_PLIC",
         swaccess: "rw",
         hwaccess: "hro",
