+++ hw/ip/rv_plic/rtl/rv_plic_reg_pkg.sv
@@ -7,9 +7,8 @@
 package rv_plic_reg_pkg;
 
   // Param list
-  localparam int IP = 32;
-  localparam int LE = 32;
-  localparam int IE0 = 32;
+  localparam int NumSrc = 32;
+  localparam int NumTarget = 1;
 
 // Register to internal design logic
 typedef struct packed {
