+++ hw/top_earlgrey/rtl/rv_plic_reg_pkg.sv
@@ -7,9 +7,8 @@
 package rv_plic_reg_pkg;
 
   // Param list
-  localparam int IP = 54;
-  localparam int LE = 54;
-  localparam int IE0 = 54;
+  localparam int NumSrc = 54;
+  localparam int NumTarget = 1;
 
 // Register to internal design logic
 typedef struct packed {
